fix consent flag and date not being updated
